Use a dedicated filesystem layout for vcpkg debug configuration
The vcpkg debug configuration is special, in the sense that the prefix is set to debug, and libraries go into `debug/bin` and `debug/lib`, but the headers go into `include` (without the `debug/` prefix), and are shared with the release configuration.  The same is true for manpages.

Most autoconfig-based projects allow you to configure these paths via `--includedir` and `--mandir`, but these arguments are ignored by gnustep-make.

This commit works around that by creating a separate filesystem layout, named vcpkg-debug, which is based on the fhs debug layout, but accounts for the exceptions mentioned above.